requirements.txt
streamlit
torch
torchvision
transformers
Pillow
numpy
matplotlib
